refactor(services): rename project directory to projects
- Rename services/project directory to services/projects to follow RESTful API naming convention
- Update all import paths to reflect the new directory structure
