+++ components/Review/Detail.tsx
@@ -1,5 +1,5 @@
 import PostDetailCard, { BasePostDetailData } from '@/shared/components/Post/PostDetailCard';
-import { ProjectReviewSchema } from '@/services/project/reviews';
+import { ProjectReviewSchema } from '@/services/projects/reviews';
 import { BaseUserSchema } from '@/services/users';
 import { CommentType } from '@/services/comments';
 import RadioGroup from './RadioGroup';
